refactor: delete shadow app/documents.py, repoint tests to documents_pkg
documents.py defined 14 router endpoints that were never mounted (server.py
mounts documents_pkg.router) and a divergent older copy of search_documents
missing Sentry breadcrumbs, result caching, ef_search tuning, and query_type
telemetry. Tests had been binding to the stale version.

Deletes 1,815 lines of dead code; tests now exercise the live implementation.
